Move Linux HAL and hardware classes to oshi-common
Move LinuxHardwareAbstractionLayer and its non-native dependencies
(LinuxComputerSystem, LinuxFirmware, LinuxBaseboard, LinuxSensors,
LinuxSoundCard) to oshi-common. Move UnixDisplay and Xrandr driver
to oshi-common. Make getPrinters() use interface default (empty list)
on common HAL, pushed down to JNA (UnixPrinter) and FFM (CupsPrinter)
subclasses. Make getGraphicsCards() abstract on common HAL due to
LinuxGraphicsCard -> LinuxGpuStats -> NvmlUtil JNA dependency chain.

Move 8 native-free Linux hardware base classes to oshi-common:
LinuxCentralProcessor, LinuxGlobalMemory, LinuxHWDiskStore,
LinuxLogicalVolumeGroup, LinuxNetworkIF, LinuxPowerSource,
LinuxUsbDevice, and LinuxVirtualMemory. LinuxCentralProcessor and
LinuxLogicalVolumeGroup made public for cross-package subclass access.
Updated imports in all JNA and FFM subclasses. Moved
LinuxPowerSourceTest to oshi-common.

LinuxGraphicsCard and LinuxGpuStats remain in oshi-core due to
NvmlUtil (JNA) dependency in createStatsSession().
