+++ oshi-common/src/main/java/oshi/hardware/common/platform/linux/LinuxSoundCard.java
@@ -2,7 +2,7 @@
  * Copyright 2018-2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
-package oshi.hardware.platform.linux;
+package oshi.hardware.common.platform.linux;
 
 import java.io.File;
 import java.util.ArrayList;
