+++ oshi-common/src/main/java/oshi/hardware/common/platform/linux/LinuxVirtualMemory.java
@@ -2,7 +2,7 @@
  * Copyright 2019-2026 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
-package oshi.hardware.platform.linux;
+package oshi.hardware.common.platform.linux;
 
 import static oshi.util.Memoizer.defaultExpiration;
 import static oshi.util.Memoizer.memoize;
