+++ oshi-common/src/main/java/oshi/hardware/common/platform/unix/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Provides common Unix hardware implementations shared by JNA and FFM modules.
+ */
+package oshi.hardware.common.platform.unix;
