+++ oshi-common/src/test/java/module-info.test
@@ -9,6 +9,8 @@
 --add-opens
   com.github.oshi.common/oshi.software.common=org.junit.platform.commons
 --add-opens
+  com.github.oshi.common/oshi.hardware.common.platform.linux=org.junit.platform.commons
+--add-opens
   com.github.oshi.common/oshi.util.driver.linux=org.junit.platform.commons
 --add-opens
   com.github.oshi.common/oshi.util.driver.linux.proc=org.junit.platform.commons
