+++ oshi-core-java25/src/main/java/oshi/hardware/platform/linux/LinuxHWDiskStoreFFM.java
@@ -20,6 +20,7 @@ import oshi.annotation.concurrent.ThreadSafe;
 import oshi.ffm.linux.UdevFunctions;
 import oshi.hardware.HWDiskStore;
 import oshi.hardware.HWPartition;
+import oshi.hardware.common.platform.linux.LinuxHWDiskStore;
 import oshi.util.Constants;
 import oshi.util.ParseUtil;
 import oshi.util.linux.DevPath;
