+++ oshi-core-java25/src/main/java/oshi/hardware/platform/linux/LinuxLogicalVolumeGroupFFM.java
@@ -22,6 +22,7 @@ import org.slf4j.LoggerFactory;
 
 import oshi.ffm.linux.UdevFunctions;
 import oshi.hardware.LogicalVolumeGroup;
+import oshi.hardware.common.platform.linux.LinuxLogicalVolumeGroup;
 import oshi.util.Util;
 import oshi.util.linux.DevPath;
