+++ oshi-core/src/main/java/oshi/hardware/platform/linux/LinuxHardwareAbstractionLayerJNA.java
@@ -9,11 +9,16 @@ import java.util.List;
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.hardware.CentralProcessor;
 import oshi.hardware.GlobalMemory;
+import oshi.hardware.GraphicsCard;
 import oshi.hardware.HWDiskStore;
 import oshi.hardware.LogicalVolumeGroup;
 import oshi.hardware.NetworkIF;
 import oshi.hardware.PowerSource;
+import oshi.hardware.Printer;
 import oshi.hardware.UsbDevice;
+import oshi.hardware.common.platform.linux.LinuxGlobalMemory;
+import oshi.hardware.common.platform.linux.LinuxHardwareAbstractionLayer;
+import oshi.hardware.platform.unix.UnixPrinter;
 import oshi.software.os.linux.LinuxOperatingSystemJNA;
 
 /**
@@ -51,6 +56,16 @@ public final class LinuxHardwareAbstractionLayerJNA extends LinuxHardwareAbstrac
     @Override
     public List<UsbDevice> getUsbDevices(boolean tree) {
         return LinuxUsbDeviceJNA.getUsbDevices(tree);
+    }
+
+    @Override
+    public List<GraphicsCard> getGraphicsCards() {
+        return LinuxGraphicsCard.getGraphicsCards();
+    }
+
+    @Override
+    public List<Printer> getPrinters() {
+        return UnixPrinter.getPrinters();
     }
 
     @Override
