+++ oshi-core/src/main/java/oshi/hardware/platform/unix/aix/AixHardwareAbstractionLayer.java
@@ -28,7 +28,7 @@ import oshi.hardware.Sensors;
 import oshi.hardware.SoundCard;
 import oshi.hardware.UsbDevice;
 import oshi.hardware.common.AbstractHardwareAbstractionLayer;
-import oshi.hardware.platform.unix.UnixDisplay;
+import oshi.hardware.common.platform.unix.UnixDisplay;
 import oshi.hardware.platform.unix.UnixPrinter;
 
 /**
