+++ oshi-core/src/main/java/oshi/hardware/platform/unix/openbsd/OpenBsdHardwareAbstractionLayer.java
@@ -20,8 +20,8 @@ import oshi.hardware.Sensors;
 import oshi.hardware.SoundCard;
 import oshi.hardware.UsbDevice;
 import oshi.hardware.common.AbstractHardwareAbstractionLayer;
+import oshi.hardware.common.platform.unix.UnixDisplay;
 import oshi.hardware.platform.unix.BsdNetworkIF;
-import oshi.hardware.platform.unix.UnixDisplay;
 import oshi.hardware.platform.unix.UnixPrinter;
 
 /**
