+++ oshi-core/src/main/java/oshi/hardware/platform/unix/solaris/SolarisHardwareAbstractionLayer.java
@@ -20,7 +20,7 @@ import oshi.hardware.Sensors;
 import oshi.hardware.SoundCard;
 import oshi.hardware.UsbDevice;
 import oshi.hardware.common.AbstractHardwareAbstractionLayer;
-import oshi.hardware.platform.unix.UnixDisplay;
+import oshi.hardware.common.platform.unix.UnixDisplay;
 import oshi.hardware.platform.unix.UnixPrinter;
 
 /**
